+++ src/platform.rs
@@ -2,7 +2,7 @@ mod device_gl;
 mod window_sdl;
 
 pub use device_gl::{
-    FrontFace, GraphicsDevice, RawFramebuffer, RawIndexBuffer, RawProgram, RawTexture,
-    RawVertexBuffer, UniformLocation,
+    GraphicsDevice, RawFramebuffer, RawIndexBuffer, RawProgram, RawTexture, RawVertexBuffer,
+    UniformLocation,
 };
 pub use window_sdl::{handle_events, Window};